Fixed rendering, soon to fix planets again now that they can be visualized along with their orbits

+++ ProjectBane/Bane-Graphics-Core/GraphicsCore/D3D12/D3D12GraphicsDevice.cpp
@@ -272,7 +272,7 @@ static std::vector<uint8> CompileFromByteCode(const std::string& InByteCode, ESH
 	}
 
 	std::vector<uint8> RetCode(static_cast<uint32>(CompiledCode->GetBufferSize()));
-	memcpy(RetCode.data(), CompiledCode->GetBufferPointer(), RetCode.size());
+	memcpy(RetCode.data(), CompiledCode->GetBufferPointer(), static_cast<uint32>(RetCode.size()));
 	CompiledCode->Release();
 	return RetCode;
 }
